Daily rainfall data set from a network of weather stations in France

A data set with precipitation amounts in mm from nine weather stations.

rainfall

Format

A data frame with 10 columns.

y.date

Date

BREST

Precipitation amounts in Brest (oceanic weather)

LANVEOC

Precipitation amounts in Lanveoc (oceanic weather)

QUIMPER

Precipitation amounts in Quimper (oceanic weather)

BORMES

Precipitation amounts in Bormes (mediterranean weather)

LE.LUC

Precipitation amounts in Le Luc (mediterranean weather)

HYERES

Precipitation amounts in Hyeres (mediterranean weather)

NANCY

Precipitation amounts in Nancy (continental weather)

METZ

Precipitation amounts in Metz (continental weather)

ROVILLE

Precipitation amounts in Roville (continental weather)

SEASON

Season

Source

Méteo France.